+++ packages/orq-rc/src/models/operations/listdatasets.ts
@@ -169,7 +169,7 @@ export const ListDatasetsData$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-06-15T08:55:55.775Z",
+    "2026-06-15T10:38:53.975Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {