+++ packages/orq-rc/RELEASES.md
@@ -9989,3 +9989,13 @@ Based on:
 - [typescript v4.11.0-rc.20] packages/orq-rc
 ### Releases
 - [NPM v4.11.0-rc.20] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.20 - packages/orq-rc
+
+## 2026-06-15 10:41:59
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.778.0 (2.904.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.11.0-rc.21] packages/orq-rc
+### Releases
+- [NPM v4.11.0-rc.21] https://www.npmjs.com/package/@orq-ai/node/v/4.11.0-rc.21 - packages/orq-rc